requirements.txt
Flask==0.10.1
itsdangerous==0.23
Jinja2==2.7.1
MarkupSafe==0.18
path.py==4.4
Werkzeug==0.9.4
